InstallTools.sh sdkman fix
Install sdkman was always called. Check sdk folder to decide whether to
install is or not.

+++ installTools.sh
@@ -1,16 +1,24 @@
 #!/bin/bash
 source basic.sh
 
+function command_exists () {
+    type "$1" &> /dev/null ;
+}
+
 function installBasics() {
-	tools=(terminator vim unzip zip)
+	tools=(terminator vim unzip zip tmux)
 
 	for i in ${tools[@]}; do
-		echo "install ${i}"
-		sudo apt-get install $i
+		if command_exists $i ; then
+        		eco "${i} is already installed"
+		else
+			eco "install ${i}"
+			sudo apt-get install $i
+		fi
 	done
 
 
-	echo "Config git lg alias"
+	eco "Config git lg alias"
 
 	git config --global alias.lg "log --color --graph --pretty=format:'%Cred%h%Creset -%C(yellow)%d%Creset %s %Cgreen(%cr) %C(bold blue)<%an>%Creset' --abbrev-commit"
 	git config --global core.editor "vim"
@@ -20,22 +28,23 @@ function installBasics() {
 
 installBasics
 
-eco "Set tmux conf"
+if command_exists ; then
+	eco "Set tmux conf"
 
-if [[ -L ~/.tmux.conf ]]; then
-	echo "-------------- .tmux does not exists"
-	mv ~/.tmux.conf ~/.tmux.conf.copy
+	if [[ -L ~/.tmux.conf ]]; then
+		echo "-------------- .tmux does not exists"
+		mv ~/.tmux.conf ~/.tmux.conf.copy
+	fi
+
+	ln -s $(pwd)/f-tmux.conf ~/.tmux.conf
 fi
 
-ln -s $(pwd)/f-tmux.conf ~/.tmux.conf
-
-
-if [ -z "$(command -v sdk)"  ]; then
+if [ -d "$HOME/.sdkman" ]  ; then
 	eco "Sdk man is already installed"
-else 
+else
 	eco "Installing Sdk Man"
 	curl -s "https://get.sdkman.io" | bash
+	source "$HOME/.sdkman/bin/sdkman-init.sh"
+	sdk version
 fi
 
-source "$HOME/.sdkman/bin/sdkman-init.sh"
-sdk version
